Split sentry file into multiple files, Fix stacktrace merging performance for > 0.46
